pyproject.toml
[tool.poetry]
name = "line-bot"
version = "0.1.0"
description = ""
authors = ["Schwannden-Kuo <Schwannden.Kuo@dell.com>"]
readme = "README.md"

[tool.mypy]
mypy_path = "./line_bot"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
fastapi = "^0.103.1"
loguru = "^0.7.1"
pydantic-settings = "^2.0.3"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
autoflake = "^2.2.1"
isort = "^5.12.0"
mypy = "^1.5.1"
types-requests = "^2.31.0.2"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
